cleanup after security fix
